Rename names.hpp to types.hpp
