+++ include/modules/fs.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "components/config.hpp"
-#include "modules/meta/names.hpp"
 #include "modules/meta/timer_module.hpp"
+#include "modules/meta/types.hpp"
 #include "settings.hpp"
 
 POLYBAR_NS
