+++ include/modules/github.hpp
@@ -2,8 +2,8 @@
 
 #include <atomic>
 
-#include "modules/meta/names.hpp"
 #include "modules/meta/timer_module.hpp"
+#include "modules/meta/types.hpp"
 #include "settings.hpp"
 #include "utils/http.hpp"
