+++ include/modules/i3.hpp
@@ -4,7 +4,7 @@
 
 #include "components/config.hpp"
 #include "modules/meta/event_module.hpp"
-#include "modules/meta/names.hpp"
+#include "modules/meta/types.hpp"
 #include "utils/i3.hpp"
 #include "utils/io.hpp"
