+++ include/modules/memory.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "modules/meta/names.hpp"
 #include "modules/meta/timer_module.hpp"
+#include "modules/meta/types.hpp"
 #include "settings.hpp"
 
 POLYBAR_NS
