+++ include/modules/mpd.hpp
@@ -4,7 +4,7 @@
 
 #include "adapters/mpd.hpp"
 #include "modules/meta/event_module.hpp"
-#include "modules/meta/names.hpp"
+#include "modules/meta/types.hpp"
 #include "utils/env.hpp"
 
 POLYBAR_NS
