+++ include/modules/script.hpp
@@ -2,7 +2,7 @@
 
 #include "adapters/script_runner.hpp"
 #include "modules/meta/base.hpp"
-#include "modules/meta/names.hpp"
+#include "modules/meta/types.hpp"
 #include "utils/command.hpp"
 #include "utils/io.hpp"
