+++ include/modules/xwindow.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "modules/meta/event_handler.hpp"
-#include "modules/meta/names.hpp"
 #include "modules/meta/static_module.hpp"
+#include "modules/meta/types.hpp"
 #include "x11/ewmh.hpp"
 #include "x11/icccm.hpp"
 #include "x11/window.hpp"
